examples/advanced_tensorflow/pyproject.toml
[build-system]
requires = [
    "poetry==1.1.6",
]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "advanced_flwr_tensorflow"
version = "0.1.0"
description = "Advanced Flower/TensorFlow Example"
authors = ["The Flower Authors <enquiries@flower.dev>"]

[tool.poetry.dependencies]
python = "^3.7.1"
flwr = "^0.16.0"  # For development: { path = "../../", develop = true }
tensorflow-cpu = "^2.4.1"
wget = "^3.2"
scikit-learn = "^0.23.2"
tensorflow = "^2.4.1"
opencv-python = "^4.5.2.52"
pandas = "^1.2.4"
google-cloud-storage = "^1.38.0"
google-cloud = "^0.34.0"
